# perf(serde_v8): smallvec ByteString