Avoiding heap allocs when deserializing ByteStrings with less than 16 chars

This comes at no memory overhead, since both `Vec<u8>` and `SmallVec<[u8; 16]>` (w/ union) are 24 bytes

Before:
```
test de_bstr_v8_1024_b    ... bench:          50 ns/iter (+/- 3)
test de_bstr_v8_12_b      ... bench:          35 ns/iter (+/- 1)
```

After:
```
test de_bstr_v8_1024_b    ... bench:          50 ns/iter (+/- 2)
test de_bstr_v8_12_b      ... bench:          15 ns/iter (+/- 0)
```

## TODO

- [ ] Fix some From/Into conversions