@andreubotella 2022-04-05
Can we assert that `Vec<u8>` and `SmallVec<[u8; 16]>` have the same size?