@AaronO 2022-04-12
> > Can we assert that Vec and SmallVec<[u8; 16]> have the same size?
> 
> `assert_eq!(sizeof::<Vec<u8>>(), sizeof::<SmallVec<[u8; 16]>>())`

They only have the same size when using `smallvec`'s [`features = ["union"]`](https://docs.rs/smallvec/latest/smallvec/#union)

> When the union feature is enabled smallvec will track its state (inline or spilled) without the use of an enum tag, reducing the size of the smallvec by one machine word. This means that there is potentially no space overhead compared to Vec. **Note that smallvec can still be larger than Vec if the inline buffer is larger than two machine words.**

This isn't a safety or correctness guarantee, just a matter of mem efficiency that's not super obscure and documented by `smallvec`